NGAE-XRD/requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
opencv-python>=4.8
pyts>=0.13
Pillow>=10.0
PyYAML>=6.0
tensorflow>=2.12
